# sdk/ts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "vmond",
    "description": "Vibemon server core: engine, image pipeline, v1 HTTP API, mesh",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/info": {
      "get": {
        "tags": [],
        "operationId": "info",
        "responses": {
          "200": {
            "description": "server info"
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/pools/{reference}": {
      "put": {
        "tags": [],
        "operationId": "pool_set",
        "parameters": [
          {
            "name": "reference",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PoolPutBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [],
        "operationId": "pool_delete",
        "parameters": [
          {
            "name": "reference",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sandboxes": {
      "get": {
        "tags": [],
        "operationId": "list_sandboxes",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxesBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "create_sandbox",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SandboxCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "created"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sandboxes/{id}/exec": {
      "get": {
        "tags": [],
        "operationId": "exec_ws",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "exec websocket"
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "exec_capture",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecCaptureBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sandboxes/{id}/files": {
      "get": {
        "tags": [],
        "operationId": "file_read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [],
        "operationId": "file_write",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [],
        "operationId": "file_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sandboxes/{id}/files/list": {
      "get": {
        "tags": [],
        "operationId": "file_list",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v1/sandboxes/{id}/files/stat": {
      "get": {
        "tags": [],
        "operationId": "file_stat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/v1/sandboxes/{id}/migrate": {
      "post": {
        "tags": [],
        "operationId": "migrate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MigrateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v1/sandboxes/{id}/network": {
      "get": {
        "tags": [],
        "operationId": "network_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [],
        "operationId": "network_set",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NetworkBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/v1/sandboxes/{id}/snapshots": {
      "post": {
        "tags": [],
        "operationId": "snapshot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v1/sandboxes/{id}/snapshots/fs": {
      "post": {
        "tags": [],
        "operationId": "snapshot_fs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotFsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/v1/sandboxes/{id}/tunnels": {
      "get": {
        "tags": [],
        "operationId": "tunnels",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/v1/volumes/{name}": {
      "put": {
        "tags": [],
        "operationId": "volume_create",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [],
        "operationId": "volume_delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ExecBody": {
        "type": "object",
        "description": "`POST /{id}/exec` body and the first WS exec frame (§2.4).",
        "properties": {
          "cmd": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cwd": {
            "type": [
              "string",
              "null"
            ],
            "description": "Legacy alias for `workdir` accepted on input."
          },
          "env": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "timeout": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "tty": {
            "type": "boolean"
          },
          "workdir": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ExecCaptureBody": {
        "type": "object",
        "required": [
          "exit",
          "stdout_b64",
          "stderr_b64"
        ],
        "properties": {
          "exit": {
            "type": "integer",
            "format": "int64"
          },
          "stderr_b64": {
            "type": "string"
          },
          "stdout_b64": {
            "type": "string"
          }
        }
      },
...
      "MigrateBody": {
        "type": "object",
        "description": "`POST /{id}/migrate` body (admin; mesh).",
        "required": [
          "target"
        ],
        "properties": {
          "target": {
            "type": "string"
          }
        }
      },
...
      "OkBody": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
      "PoolPutBody": {
        "type": "object",
        "description": "`PUT /v1/pools/{ref}` body (= today's `prewarm`).",
        "required": [
          "size"
        ],
        "properties": {
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": {
          "description": "Template kwargs forwarded to the image pipeline."
        }
      },
...
      "SandboxCreate": {
        "type": "object",
        "description": "`POST /v1/sandboxes` body: today's `SandboxCreate` field set plus\n`command` (foreground entrypoint override composed by the CLI).\nExternal `remote_page_*` fields are rejected by validation.",
        "properties": {
          "arch": {
            "type": [
              "string",
              "null"
            ]
          },
          "block_network": {
            "type": "boolean"
          },
          "command": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Foreground entrypoint override (`command?: [str]`, new in v1)."
          },
          "context": {
            "type": "string"
          },
          "cpus": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "disk_mb": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "dockerfile": {
            "type": [
              "string",
              "null"
            ]
          },
          "egress_allow": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "egress_allow_domains": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "env": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "fs_dir": {
            "type": [
              "string",
              "null"
            ]
          },
          "ha": {
            "type": [
              "string",
              "null"
            ]
          },
          "idempotency_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "image": {
            "type": [
              "string",
              "null"
            ]
          },
          "inbound_cidr_allowlist": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "memory": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "pool_size": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "ports": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          "readiness_probe": {},
          "remote_page_digest": {
            "type": [
              "string",
              "null"
            ]
          },
          "remote_page_token": {
            "type": [
              "string",
              "null"
            ]
          },
          "remote_page_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Rejected for external callers (mesh-internal restore plumbing)."
          },
          "secrets": {
            "type": [
              "array",
              "null"
            ],
            "items": {},
            "description": "Request-scoped secrets: `[{name, values}]`; never persisted."
          },
          "tags": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "template": {
            "type": [
              "string",
              "null"
            ]
          },
          "timeout": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "timeout_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "volumes": {
            "type": [
              "object",
              "null"
            ],
            "description": "Mountpoint -> volume name or `{name, read_only}`.",
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            }
          },
          "workdir": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
      "SandboxesBody": {
        "type": "object",
        "required": [
          "sandboxes"
        ],
        "properties": {
          "sandboxes": {
            "type": "array",
            "items": {}
          }
        }
      },
...
    }
  },
...
}